// workspaces/openshift-image-registry/plugins/openshift-image-registry/src/types.ts
/*
 * Copyright Red Hat, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type ImageStream = {
  uid: string;
  name: string;
  namespace: string;
  last_modified: string;
  tags: string[];
  dockerImageRepo: string;
};

export type ImageStreamMetadata = ImageStream & {
  description?: string;
  version?: string;
  size?: string;
};